requirements.txt
Pillow==8.4.0
protobuf==3.19.1
PyYAML==6.0
